fix for npc trader console errors with "argument -1 out of range"

+++ src/npc.cpp
@@ -834,8 +834,8 @@ int NpcScriptInterface::luaOpenShopWindow(lua_State* L)
 
 		item.specialId = getField<uint32_t>(L, tableIndex, "specialId");
 
-		item.buyPrice = getField<uint32_t>(L, tableIndex, "buy");
-		item.sellPrice = getField<uint32_t>(L, tableIndex, "sell");
+		item.buyPrice = getField<int64_t>(L, tableIndex, "buy");
+		item.sellPrice = getField<int64_t>(L, tableIndex, "sell");
 		item.realName = getFieldString(L, tableIndex, "name");
 		item.funcShop = getField<uint32_t>(L, tableIndex, "funcShop");
 
@@ -1054,8 +1054,8 @@ int NpcScriptInterface::luaNpcOpenShopWindow(lua_State* L)
 			lua_pop(L, 1);
 		}
 
-		item.buyPrice = getField<uint32_t>(L, tableIndex, "buy");
-		item.sellPrice = getField<uint32_t>(L, tableIndex, "sell");
+		item.buyPrice = getField<int64_t>(L, tableIndex, "buy");
+		item.sellPrice = getField<int64_t>(L, tableIndex, "sell");
 		item.realName = getFieldString(L, tableIndex, "name");
 
 		items.push_back(item);
